Fix/correciones lautaro

+++ src/dao/PlatoDao.java
@@ -62,7 +62,7 @@ public class PlatoDao {
 		}
 	}
 
-	public Plato traer(long idPlato) {
+	public Plato traer(int idPlato) {
 		Plato objeto = null;
 		try {
 			iniciaOperacion();
